removed a warning
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk/wired@588 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/gui/MainWindow.cpp
@@ -1277,7 +1277,7 @@ void					MainWindow::OnSettings(wxCommandEvent &event)
 	{
       FileConverter->SetFormat((PaSampleFormat)Audio->SampleFormat);
       FileConverter->SetBufferSize(Audio->SamplesPerBuffer);		
-      FileConverter->SetSampleRate(Audio->SampleRate);		
+      FileConverter->SetSampleRate((long unsigned int)Audio->SampleRate);		
 	  Audio->OpenStream();
 	  if (Audio->StartStream())
 	    AudioMutex.Unlock();/* This will unlock the sequencer	\
